@@ -1,56 +1,53 @@
 {
   "name": "react-bootsrap",
-  "version": "1.1.0",
-  "description": "A foundation for rapid React prototypes",
+  "version": "0.0.1",
+  "description": "React frontend for the Sump Thing project",
   "main": "index.js",
   "dependencies": {
-    "lodash": "^4.17.4",
-    "react": "^16.0.0"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-class-properties": "^7.0.0",
-    "@babel/plugin-proposal-decorators": "^7.0.0",
-    "@babel/plugin-transform-runtime": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
+    "@babel/core": "^7.4.4",
+    "@babel/plugin-proposal-class-properties": "^7.4.4",
+    "@babel/plugin-proposal-decorators": "^7.4.4",
+    "@babel/plugin-transform-runtime": "^7.4.4",
+    "@babel/preset-env": "^7.4.4",
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-stage-2": "^7.0.0",
-    "@material-ui/core": "^3.0.2",
-    "@material-ui/icons": "^3.0.1",
+    "@material-ui/core": "^4.0.0-beta.1",
+    "@material-ui/icons": "^3.0.2",
+    "@material-ui/styles": "^3.0.0-alpha.10",
+    "@material-ui/system": "^3.0.0-alpha.2",
+    "add": "^2.0.6",
     "babel-core": "^7.0.0-0",
-    "babel-eslint": "^9.0.0",
-    "babel-jest": "^23.4.2",
-    "babel-loader": "^8.0.2",
+    "babel-eslint": "^10.0.1",
+    "babel-jest": "^24.8.0",
+    "babel-loader": "^8.0.6",
     "babel-runtime": "^6.22.0",
-    "css-loader": "0.26.1",
-    "enzyme": "^3.1.0",
-    "enzyme-adapter-react-16": "^1.0.2",
-    "eslint": "^4.3.0",
-    "eslint-config-airbnb": "^15.1.0",
-    "eslint-plugin-babel": "^4.1.2",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jest": "^20.0.3",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.1.0",
-    "file-loader": "^0.10.0",
-    "happypack": "^5.0.0",
+    "css-loader": "2.1.1",
+    "enzyme": "^3.9.0",
+    "enzyme-adapter-react-16": "^1.13.0",
+    "eslint": "^5.16.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-babel": "^5.3.0",
+    "eslint-plugin-import": "^2.17.2",
+    "eslint-plugin-jest": "^22.5.1",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-react": "^7.13.0",
+    "file-loader": "^3.0.1",
+    "happypack": "^5.0.1",
     "html-webpack-plugin": "^3.2.0",
-    "jest": "^23.5.0",
-    "prop-types": "^15.6.0",
+    "react": "^16.8.6",
     "react-addons-test-utils": "^15.6.0",
-    "react-dom": "^16.0.0",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "react-router-dom": "^4.2.2",
-    "react-test-renderer": "^15.6.1",
-    "url-loader": "1.1.1",
-    "webpack": "^4.17.2",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-server": "^3.1.8"
+    "react-dom": "^16.8.6",
+    "react-hot-loader": "^4.8.4",
+    "react-router-dom": "^5.0.0",
+    "react-test-renderer": "^16.8.6",
+    "url-loader": "1.1.2",
+    "webpack": "^4.31.0",
+    "webpack-cli": "^3.3.2",
+    "webpack-dev-server": "^3.3.1",
+    "yarn": "^1.16.0"
   },
   "scripts": {
-    "start": "docker-compose up",
-    "stop": "docker-compose down",
-    "webpack-dev-server": "webpack-dev-server --config config/webpack/main.js --colors",
+    "dev": "webpack-dev-server --config config/webpack/main.js --colors",
     "test": "./node_modules/.bin/jest --config='./jest.config.js'",
     "test:watch": "./node_modules/.bin/jest --watch --config='./jest.config.js'",
     "build": "webpack --config config/webpack/production.js"
@@ -68,7 +65,7 @@
       [
         "@babel/plugin-proposal-decorators",
         {
-          "legacy": true
+          "decoratorsBeforeExport": false
         }
       ]
     ],
@@ -79,8 +76,8 @@
   },
   "repository": {
     "type": "git",
-    "url": "ssh://developertown@developertown.visualstudio.com:22/DefaultCollection/DT%20Starts/_git/react-bootsrap"
+    "url": "https://github.com/pyazo/sump-thing-frontend"
   },
-  "author": "DT Starts",
+  "author": "Pyazo, LLC",
   "license": "ISC"
 }
